Refactor to top-level phs folder
